- improve node and interface of visualea - add mode disabled to curveeditor
git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/vplants/vplants/trunk/PlantGL@8260 ab253cce-29fb-0310-bb2f-979600cdbdeb

+++ src/wrapper/scenegraph/export_nurbscurve.cpp
@@ -65,7 +65,7 @@ std::string nc_repr( NurbsCurve* p )
   if (!p->isStrideToDefault())
 	  ss << ", stride = " << p->getStride();
   if (!p->isWidthToDefault())
-	  ss << ", width = " << p->getWidth();
+	  ss << ", width = " << (int)p->getWidth();
   ss << ")";
   return ss.str();
 }
@@ -141,7 +141,7 @@ std::string nc2_repr( NurbsCurve2D* p )
   if (!p->isStrideToDefault())
 	  ss << ", stride = " << p->getStride();
   if (!p->isWidthToDefault())
-	  ss << ", width = " << p->getWidth();
+	  ss << ", width = " << (int)p->getWidth();
   ss << ")";
   return ss.str();
 }
